backend/requirements.txt
Flask==2.0.0
python-dotenv==0.21.1
s3fs==2023.1.0
openpyxl==3.1.1
pandas==1.3.5
gunicorn
